+++ bin/proving-service/src/api/mod.rs
@@ -1,8 +1,11 @@
-use miden_objects::transaction::TransactionWitness;
+use miden_objects::{
+    batch::ProposedBatch, transaction::TransactionWitness, MIN_PROOF_SECURITY_LEVEL,
+};
 use miden_tx::{
     utils::{Deserializable, Serializable},
     LocalTransactionProver, TransactionProver,
 };
+use miden_tx_batch_prover::LocalBatchProver;
 use tokio::{net::TcpListener, sync::Mutex};
 use tonic::{Request, Response, Status};
 use tracing::instrument;
@@ -10,7 +13,7 @@ use tracing::instrument;
 use crate::{
     generated::{
         api_server::{Api as ProverApi, ApiServer},
-        ProveTransactionRequest, ProveTransactionResponse,
+        ProveRequest, ProveResponse,
     },
     utils::MIDEN_PROVING_SERVICE,
 };
@@ -21,48 +24,118 @@ pub struct RpcListener {
 }
 
 impl RpcListener {
-    pub fn new(listener: TcpListener) -> Self {
-        let api_service = ApiServer::new(ProverRpcApi::default());
+    pub fn new(listener: TcpListener, is_tx_prover: bool, is_batch_prover: bool) -> Self {
+        let prover_rpc_api = ProverRpcApi::new(is_tx_prover, is_batch_prover);
+        let api_service = ApiServer::new(prover_rpc_api);
         Self { listener, api_service }
     }
 }
 
-#[derive(Default)]
 pub struct ProverRpcApi {
-    local_prover: Mutex<LocalTransactionProver>,
+    tx_prover: Option<Mutex<LocalTransactionProver>>,
+    batch_prover: Option<Mutex<LocalBatchProver>>,
+}
+
+impl ProverRpcApi {
+    pub fn new(is_tx_prover: bool, is_batch_prover: bool) -> Self {
+        let tx_prover = if is_tx_prover {
+            Some(Mutex::new(LocalTransactionProver::default()))
+        } else {
+            None
+        };
+
+        let batch_prover = if is_batch_prover {
+            Some(Mutex::new(LocalBatchProver::new(MIN_PROOF_SECURITY_LEVEL)))
+        } else {
+            None
+        };
+
+        Self { tx_prover, batch_prover }
+    }
+    #[instrument(
+        target = MIDEN_PROVING_SERVICE,
+        name = "remote_prover:prove_tx",
+        skip_all,
+        ret(level = "debug"),
+        fields(id = tracing::field::Empty),
+        err
+    )]
+    pub fn prove_tx(
+        &self,
+        request: Request<ProveRequest>,
+    ) -> Result<Response<ProveResponse>, tonic::Status> {
+        let tx_prover = self
+            .tx_prover
+            .as_ref()
+            .ok_or(Status::unimplemented("Transaction prover is not enabled"))?;
+        let prover = tx_prover
+            .try_lock()
+            .map_err(|_| Status::resource_exhausted("Server is busy handling another request"))?;
+
+        let transaction_witness = TransactionWitness::read_from_bytes(&request.get_ref().payload)
+            .map_err(invalid_argument)?;
+
+        let proof = prover.prove(transaction_witness).map_err(internal_error)?;
+
+        // Record the transaction_id in the current tracing span
+        let transaction_id = proof.id();
+        tracing::Span::current().record("id", tracing::field::display(&transaction_id));
+
+        Ok(Response::new(ProveResponse { payload: proof.to_bytes() }))
+    }
+
+    #[instrument(
+        target = MIDEN_PROVING_SERVICE,
+        name = "remote_prover:prove_batch",
+        skip_all,
+        ret(level = "debug"),
+        fields(id = tracing::field::Empty),
+        err
+    )]
+    pub fn prove_batch(
+        &self,
+        request: Request<ProveRequest>,
+    ) -> Result<Response<ProveResponse>, tonic::Status> {
+        let batch_prover = self
+            .batch_prover
+            .as_ref()
+            .ok_or(Status::unimplemented("Batch prover is not enabled"))?;
+        let prover = batch_prover
+            .try_lock()
+            .map_err(|_| Status::resource_exhausted("Server is busy handling another request"))?;
+
+        let batch =
+            ProposedBatch::read_from_bytes(&request.get_ref().payload).map_err(invalid_argument)?;
+
+        let proof = prover.prove(batch).map_err(internal_error)?;
+
+        // Record the batch_id in the current tracing span
+        let batch_id = proof.id();
+        tracing::Span::current().record("id", tracing::field::display(&batch_id));
+
+        Ok(Response::new(ProveResponse { payload: proof.to_bytes() }))
+    }
 }
 
 #[async_trait::async_trait]
 impl ProverApi for ProverRpcApi {
     #[instrument(
         target = MIDEN_PROVING_SERVICE,
-        name = "prover:prove_transaction",
+        name = "remote_prover:prove",
         skip_all,
         ret(level = "debug"),
-        fields(transaction_id = tracing::field::Empty),
+        fields(id = tracing::field::Empty),
         err
     )]
-    async fn prove_transaction(
+    async fn prove(
         &self,
-        request: Request<ProveTransactionRequest>,
-    ) -> Result<Response<ProveTransactionResponse>, tonic::Status> {
-        // Try to acquire a permit without waiting
-        let prover = self
-            .local_prover
-            .try_lock()
-            .map_err(|_| Status::resource_exhausted("Server is busy handling another request"))?;
-
-        let transaction_witness =
-            TransactionWitness::read_from_bytes(&request.get_ref().transaction_witness)
-                .map_err(invalid_argument)?;
-
-        let proof = prover.prove(transaction_witness).map_err(internal_error)?;
-
-        // Record the transaction_id in the current tracing span
-        let transaction_id = proof.id();
-        tracing::Span::current().record("transaction_id", tracing::field::display(&transaction_id));
-
-        Ok(Response::new(ProveTransactionResponse { proven_transaction: proof.to_bytes() }))
+        request: Request<ProveRequest>,
+    ) -> Result<Response<ProveResponse>, tonic::Status> {
+        match request.get_ref().proof_type {
+            0 => self.prove_tx(request),
+            1 => self.prove_batch(request),
+            _ => Err(internal_error("Invalid proof type")),
+        }
     }
 }