+++ crates/miden-objects/src/batch/note_tree.rs
@@ -76,7 +76,7 @@ impl Serializable for BatchNoteTree {
 impl Deserializable for BatchNoteTree {
     fn read_from<R: ByteReader>(source: &mut R) -> Result<Self, DeserializationError> {
         let leaves = Vec::read_from(source)?;
-        let smt = SimpleSmt::with_contiguous_leaves(leaves.into_iter()).map_err(|err| {
+        let smt = SimpleSmt::with_leaves(leaves.into_iter()).map_err(|err| {
             DeserializationError::UnknownError(format!(
                 "failed to deserialize BatchNoteTree: {err}"
             ))