+++ crates/miden-proving-service-client/src/remote_prover/tx_prover.rs
@@ -1,15 +1,13 @@
-pub mod generated;
-
 use alloc::{
     boxed::Box,
     string::{String, ToString},
 };
 
-use generated::api_client::ApiClient;
 use miden_objects::transaction::{ProvenTransaction, TransactionWitness};
 use miden_tx::{utils::sync::RwLock, TransactionProver, TransactionProverError};
 
-use crate::RemoteProverError;
+use super::generated::api_client::ApiClient;
+use crate::{remote_prover::generated, RemoteProverError};
 
 // REMOTE TRANSACTION PROVER
 // ================================================================================================
@@ -87,11 +85,9 @@ impl TransactionProver for RemoteTransactionProver {
             .ok_or_else(|| TransactionProverError::other("client should be connected"))?
             .clone();
 
-        let request = tonic::Request::new(generated::ProveTransactionRequest {
-            transaction_witness: tx_witness.to_bytes(),
-        });
+        let request = tonic::Request::new(tx_witness.into());
 
-        let response = client.prove_transaction(request).await.map_err(|err| {
+        let response = client.prove(request).await.map_err(|err| {
             TransactionProverError::other_with_source("failed to prove transaction", err)
         })?;